+++ include/spatialops/structured/FVStaggeredInterpolant.h
@@ -90,7 +90,7 @@ namespace structured{
     /**
      *  @brief Obtain the set of column indices corresponding to ghost entries.
      */
-    void get_ghost_cols( std::set<int>& ghostCols ) const
+    void get_ghost_cols( std::set<size_t>& ghostCols ) const
     {
       ghostCols = get_ghost_set<SrcFieldT>( dim_, hasPlusXSideFaces_, hasPlusYSideFaces_, hasPlusZSideFaces_ );
     }
@@ -98,7 +98,7 @@ namespace structured{
     /**
      *  @brief Obtain the set of row indices corresponding to ghost entries.
      */
-    void get_ghost_rows( std::set<int>& ghostRows ) const
+    void get_ghost_rows( std::set<size_t>& ghostRows ) const
     {
       ghostRows = get_ghost_set<DestFieldT>( dim_, hasPlusXSideFaces_, hasPlusYSideFaces_, hasPlusZSideFaces_ );
     }
